+++ packages/nimble-components/src/dialog/tests/dialog.spec.ts
@@ -1,7 +1,6 @@
-import { DesignSystem } from '@microsoft/fast-foundation';
 import { html } from '@microsoft/fast-element';
 import { fixture, Fixture } from '../../utilities/tests/fixture';
-import { Dialog, ExtendedDialog, UserDismissed } from '..';
+import { Dialog, dialogTag, ExtendedDialog, UserDismissed } from '..';
 import { waitForUpdatesAsync } from '../../testing/async-helpers';
 
 // eslint-disable-next-line @typescript-eslint/no-invalid-void-type
@@ -26,8 +25,8 @@ describe('Dialog', () => {
         ) as ExtendedDialog;
     }
 
-    it('should have its tag returned by tagFor(Dialog)', () => {
-        expect(DesignSystem.tagFor(Dialog)).toBe('nimble-dialog');
+    it('should export its tag', () => {
+        expect(dialogTag).toBe('nimble-dialog');
     });
 
     it('can construct an element instance', () => {