+++ packages/nimble-components/src/tabs/tests/tabs.spec.ts
@@ -1,12 +1,8 @@
-import {
-    DesignSystem,
-    Tabs as FoundationTabs
-} from '@microsoft/fast-foundation';
-import { Tabs } from '..';
+import { Tabs, tabsTag } from '..';
 
 describe('Tabs', () => {
-    it('should have its tag returned by tagFor(FoundationTabs)', () => {
-        expect(DesignSystem.tagFor(FoundationTabs)).toBe('nimble-tabs');
+    it('should export its tag', () => {
+        expect(tabsTag).toBe('nimble-tabs');
     });
 
     it('can construct an element instance', () => {