+++ packages/nimble-components/src/tooltip/tests/tooltip.spec.ts
@@ -1,12 +1,8 @@
-import {
-    DesignSystem,
-    TooltipPosition,
-    Tooltip as FoundationTooltip
-} from '@microsoft/fast-foundation';
+import { TooltipPosition } from '@microsoft/fast-foundation';
 import { html } from '@microsoft/fast-element';
 import { fixture, Fixture } from '../../utilities/tests/fixture';
-import { Tooltip } from '..';
-import { AnchoredRegion } from '../../anchored-region';
+import { Tooltip, tooltipTag } from '..';
+import { anchoredRegionTag } from '../../anchored-region';
 import { waitForUpdatesAsync } from '../../testing/async-helpers';
 
 async function setup(): Promise<Fixture<Tooltip>> {
@@ -53,11 +49,7 @@ describe('Tooltip', () => {
         await waitForUpdatesAsync();
 
         expect(element.visible).toBeUndefined();
-        expect(
-            element.shadowRoot?.querySelector(
-                DesignSystem.tagFor(AnchoredRegion)
-            )
-        ).toBeNull();
+        expect(element.shadowRoot?.querySelector(anchoredRegionTag)).toBeNull();
 
         await disconnect();
     });
@@ -71,9 +63,7 @@ describe('Tooltip', () => {
 
         expect(element.visible).toBe(true);
         expect(
-            element.shadowRoot?.querySelector(
-                DesignSystem.tagFor(AnchoredRegion)
-            )
+            element.shadowRoot?.querySelector(anchoredRegionTag)
         ).not.toBeNull();
 
         await disconnect();
@@ -87,11 +77,7 @@ describe('Tooltip', () => {
         await waitForUpdatesAsync();
 
         expect(element.visible).toBe(false);
-        expect(
-            element.shadowRoot?.querySelector(
-                DesignSystem.tagFor(AnchoredRegion)
-            )
-        ).toBeNull();
+        expect(element.shadowRoot?.querySelector(anchoredRegionTag)).toBeNull();
 
         await disconnect();
     });
@@ -371,8 +357,8 @@ describe('Tooltip', () => {
 
     // end of position tests ^
 
-    it('should have its tag returned by tagFor(FoundationTooltip)', () => {
-        expect(DesignSystem.tagFor(FoundationTooltip)).toBe('nimble-tooltip');
+    it('should export its tag', () => {
+        expect(tooltipTag).toBe('nimble-tooltip');
     });
 
     it('can construct an element instance', () => {